% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{powerlaw_cdf_cpp}
\alias{powerlaw_cdf_cpp}
\title{Discrete power-law cumulative distribution function}
\usage{
powerlaw_cdf_cpp(x, alpha, x_min = 1L)
}
\arguments{
\item{x}{Integer vector of values at which to evaluate the CDF.}

\item{alpha}{Power-law exponent (alpha > 1).}

\item{x_min}{Lower cut-off (positive integer, default 1).}
}
\value{
Numeric vector of survival-function values.
}
\description{
Returns the survival function \eqn{P(X \geq x)} for a discrete power-law
distribution with the given parameters.
}
\examples{
powerlaw_cdf(1:10, alpha = 2.5, x_min = 1)

}
