% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{powerlaw_generate_cpp}
\alias{powerlaw_generate_cpp}
\title{Generate random samples from a discrete power-law distribution}
\usage{
powerlaw_generate_cpp(n, alpha, x_min = 1L, x_max = NULL)
}
\arguments{
\item{n}{Number of samples to generate.}

\item{alpha}{Power-law exponent (alpha > 1).}

\item{x_min}{Lower cut-off (positive integer, default 1).}

\item{x_max}{Upper cut-off. If \code{NA} (default), uses \code{x_min * 1000}.}
}
\value{
Integer vector of length \code{n}.
}
\description{
Generate random samples from a discrete power-law distribution
}
\examples{
set.seed(1)
samples <- powerlaw_generate(100, alpha = 2.5, x_min = 1)
hist(samples, breaks = 20)

}
