% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{powerlaw_pdf_cpp}
\alias{powerlaw_pdf_cpp}
\title{Discrete power-law probability density function}
\usage{
powerlaw_pdf_cpp(x, alpha, x_min = 1L)
}
\arguments{
\item{x}{Integer vector of values at which to evaluate the PMF.}

\item{alpha}{Power-law exponent (alpha > 1).}

\item{x_min}{Lower cut-off (positive integer, default 1).}
}
\value{
Numeric vector of probability mass values.
}
\description{
Returns P(X = x) for a discrete power-law distribution with the given
parameters.
}
\examples{
powerlaw_pdf(1:10, alpha = 2.5, x_min = 1)

}
